VMware engineering team are actively investigating the issue and working towards
To assist in identifying VMs that might be problematic, the below is a PowerCLI One-Liner that you can use to determine if you are at risk. Please note that this code does not check the exact Windows build called out in the KB but it gets as close as possible with the data easily exposed.